Splitup into files
Goal is to not change the external structure and exports, so all this should stay the same.
